App settings tree items base classes
