+++ appservice/src/tree/AppSettingsTreeItem.ts
@@ -4,15 +4,15 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { StringDictionary } from 'azure-arm-website/lib/models';
-import { AzureParentTreeItem, AzureTreeItem, IActionContext, ICreateChildImplContext, TreeItemIconPath } from 'vscode-azureextensionui';
+import { AzExtTreeItem, AzureParentTreeItem, ICreateChildImplContext } from 'vscode-azureextensionui';
 import { ext } from '../extensionVariables';
-import { SiteClient } from "../SiteClient";
+import { SiteClient } from '../SiteClient';
+import { AppSettingsTreeItemBase } from './AppSettingsTreeItemBase';
 import { AppSettingTreeItem } from './AppSettingTreeItem';
-import { getThemedIconPath } from './IconPath';
 import { ISiteTreeRoot } from './ISiteTreeRoot';
 
-export function validateAppSettingKey(settings: StringDictionary, client: SiteClient, newKey: string, oldKey?: string): string | undefined {
-    if (client.isLinux && /[^\w\.]+/.test(newKey)) {
+export function validateAppSettingKey(settings: StringDictionary, isLinux: boolean, newKey: string, oldKey?: string): string | undefined {
+    if (isLinux && /[^\w\.]+/.test(newKey)) {
         return 'App setting names can only contain letters, numbers (0-9), periods ("."), and underscores ("_")';
     }
 
@@ -33,27 +33,16 @@ export function validateAppSettingKey(settings: StringDictionary, client: SiteCl
     return undefined;
 }
 
-export class AppSettingsTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
-    public static contextValue: string = 'applicationSettings';
-    public readonly label: string = 'Application Settings';
-    public readonly childTypeLabel: string = 'App Setting';
-    public readonly contextValue: string = AppSettingsTreeItem.contextValue;
-    private _settings: StringDictionary | undefined;
+export class AppSettingsTreeItem extends AppSettingsTreeItemBase {
 
-    public get id(): string {
-        return 'configuration';
+    public readonly _client: SiteClient;
+
+    constructor(parent: AzureParentTreeItem<ISiteTreeRoot>) {
+        super(parent, parent.root.client);
     }
 
-    public get iconPath(): TreeItemIconPath {
-        return getThemedIconPath('settings');
-    }
-
-    public hasMoreChildrenImpl(): boolean {
-        return false;
-    }
-
-    public async loadMoreChildrenImpl(_clearCache: boolean): Promise<AzureTreeItem<ISiteTreeRoot>[]> {
-        this._settings = await this.root.client.listApplicationSettings();
+    public async loadMoreChildrenImpl(_clearCache: boolean): Promise<AzExtTreeItem[]> {
+        this._settings = await this._client.listApplicationSettings();
         const treeItems: AppSettingTreeItem[] = [];
         // tslint:disable-next-line:strict-boolean-expressions
         const properties: { [name: string]: string } = this._settings.properties || {};
@@ -65,39 +54,13 @@ export class AppSettingsTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
         return treeItems;
     }
 
-    public async editSettingItem(oldKey: string, newKey: string, value: string, context: IActionContext): Promise<void> {
-        // make a deep copy so settings are not cached if there's a failure
-        // tslint:disable-next-line: no-unsafe-any
-        const settings: StringDictionary = JSON.parse(JSON.stringify(await this.ensureSettings(context)));
-        if (settings.properties) {
-            if (oldKey !== newKey) {
-                delete settings.properties[oldKey];
-            }
-            settings.properties[newKey] = value;
-        }
-
-        this._settings = await this.root.client.updateApplicationSettings(settings);
-    }
-
-    public async deleteSettingItem(key: string, context: IActionContext): Promise<void> {
-        // make a deep copy so settings are not cached if there's a failure
-        // tslint:disable-next-line: no-unsafe-any
-        const settings: StringDictionary = JSON.parse(JSON.stringify(await this.ensureSettings(context)));
-
-        if (settings.properties) {
-            delete settings.properties[key];
-        }
-
-        this._settings = await this.root.client.updateApplicationSettings(settings);
-    }
-
-    public async createChildImpl(context: ICreateChildImplContext): Promise<AzureTreeItem<ISiteTreeRoot>> {
+    public async createChildImpl(context: ICreateChildImplContext): Promise<AppSettingTreeItem> {
         // make a deep copy so settings are not cached if there's a failure
         // tslint:disable-next-line: no-unsafe-any
         const settings: StringDictionary = JSON.parse(JSON.stringify(await this.ensureSettings(context)));
         const newKey: string = await ext.ui.showInputBox({
             prompt: 'Enter new setting key',
-            validateInput: (v: string): string | undefined => validateAppSettingKey(settings, this.root.client, v)
+            validateInput: (v: string): string | undefined => validateAppSettingKey(settings, this._client.isLinux, v)
         });
 
         const newValue: string = await ext.ui.showInputBox({
@@ -111,16 +74,8 @@ export class AppSettingsTreeItem extends AzureParentTreeItem<ISiteTreeRoot> {
         context.showCreatingTreeItem(newKey);
         settings.properties[newKey] = newValue;
 
-        this._settings = await this.root.client.updateApplicationSettings(settings);
+        this._settings = await this._client.updateApplicationSettings(settings);
 
         return await AppSettingTreeItem.createAppSettingTreeItem(this, newKey, newValue);
-    }
-
-    public async ensureSettings(context: IActionContext): Promise<StringDictionary> {
-        if (!this._settings) {
-            await this.getCachedChildren(context);
-        }
-
-        return <StringDictionary>this._settings;
     }
 }
